gcp_video_processor/requirements.txt
# requirements.txt
opencv-python-headless
numpy
moviepy==1.0.3
tqdm
google-cloud-storage